requirements.txt
markdown>=3.6
PyYAML>=6.0
feedparser>=6.0
requests>=2.31
python-dateutil>=2.9
filelock>=3.14
pytest>=8.0
